chore: reduce release PRs from 2 to 1 (#13878)
