+++ tools/release/02_publish_crates.ts
@@ -4,6 +4,7 @@ import { DenoWorkspace } from "./deno_workspace.ts";
 import { getCratesPublishOrder } from "./deps.ts";
 
 const workspace = await DenoWorkspace.load();
+const cliCrate = workspace.getCliCrate();
 
 const dependencyCrates = getCratesPublishOrder(workspace.getDependencyCrates());
 
@@ -12,6 +13,8 @@ try {
     await crate.publish();
     console.log(`Published ${i + 1} of ${dependencyCrates.length} crates.`);
   }
+
+  await cliCrate.publish();
 } finally {
   // system beep to notify error or completion
   console.log("\x07");
