+++ tools/release/deps.ts
@@ -1,3 +1,3 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-export * from "https://raw.githubusercontent.com/denoland/automation/0.3.0/mod.ts";
+export * from "https://raw.githubusercontent.com/denoland/automation/0.4.0/mod.ts";
